AI_Employee_Vault/_scripts/requirements.txt
# Python dependencies for AI Employee Watchers
# Install with: pip install -r requirements.txt

# Core dependencies
watchdog>=3.0.0

# Gmail Watcher dependencies (optional - only if using Gmail)
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0

# Optional: Playwright for WhatsApp Web automation (Silver tier)
# playwright>=1.40.0
# Run after install: playwright install

# Development dependencies
python-dotenv>=1.0.0
